src/resume_parser/parser.py
# import os
# import io
# import json
# import re
# from dotenv import load_dotenv
# from pdfminer.high_level import extract_text as pdf_extract
# import docx
# import spacy

# load_dotenv()
# nlp = spacy.load("en_core_web_sm")

# def load_skills_for_profession(profession: str) -> list[str]:
#     fname = profession.lower().replace(" ", "_") + ".json"
#     path  = os.path.join("data", "skills", fname)
#     if os.path.exists(path):
#         return json.load(open(path))
#     return []

# def read_resume(file_bytes: bytes, file_type: str) -> str:
#     if file_type == "pdf":
#         return pdf_extract(io.BytesIO(file_bytes))
#     elif file_type == "docx":
#         doc = docx.Document(io.BytesIO(file_bytes))
#         return "\n".join(p.text for p in doc.paragraphs)
#     else:
#         return file_bytes.decode("utf-8", errors="ignore")

# def parse_resume(text: str, profession: str) -> dict:
#     skills_db = load_skills_for_profession(profession)
#     doc = nlp(text)

#     # 1. Education
#     education = [
#         ent.text for ent in doc.ents
#         if ent.label_ == "ORG" and
#            any(kw in ent.text.lower() for kw in ["university","college","institute"])
#     ]
#     # 2. Years of experience
#     exps = re.findall(r"(\d+)\+?\s*years", text.lower())
#     experience_years = max(map(int, exps)) if exps else 0
#     # 3. Skills
#     skills = [
#         skill for skill in skills_db
#         if re.search(rf"\b{re.escape(skill)}\b", text, re.IGNORECASE)
#     ]

#     return {
#         "education":        education,
#         "experience_years": experience_years,
#         "skills":           skills
#     }

# src/resume_parser/parser.py

import os

try:
    import PyPDF2
except ImportError:
    PyPDF2 = None

try:
    import docx
except ImportError:
    docx = None


def parse_resume_text(path: str) -> str:
    """
    Given a path to a resume file (.txt, .pdf, .docx), extract and return its text.
    Falls back to a stub string if extraction fails.
    """
    ext = os.path.splitext(path)[1].lower()
    text = ""

    if ext == ".txt":
        # Plain text resume
        try:
            with open(path, "r", encoding="utf-8") as f:
                text = f.read()
        except Exception:
            text = ""
    
    elif ext == ".pdf" and PyPDF2 is not None:
        # PDF resume
        try:
            with open(path, "rb") as f:
                reader = PyPDF2.PdfReader(f)
                for page in reader.pages:
                    # Some pages might return None
                    page_text = page.extract_text()
                    if page_text:
                        text += page_text + "\n"
        except Exception:
            text = ""
    
    elif ext == ".docx" and docx is not None:
        # DOCX resume
        try:
            doc = docx.Document(path)
            for para in doc.paragraphs:
                text += para.text + "\n"
        except Exception:
            text = ""
    
    # If we got nothing, return a stub so the demo still runs
    if not text.strip():
        return (
            "Python pandas scikit-learn machine learning API design cloud "
            "data analysis software development"
        )
    
    return text

